Beta v3.7.1
Maintenance. The toolchain and eight dependencies move, two of them across a
major, and nothing in the app changes on purpose.

  agp 9.3.2 -> 9.4.0, gradle 9.7.0 -> 9.7.1, kotlin 2.2.10 -> 2.4.10, composeBom
  2026.06.01 -> 2026.08.00, media3 1.10.1 -> 1.11.0, appcompat 1.7.1 -> 1.8.0,
  playServicesAuth 21.4.0 -> 22.0.0, okhttp 4.12.0 -> 5.5.0, maplibre 13.3.1 ->
  13.6.0, ui-test-manifest 1.11.4 -> 1.12.0.

Note that `kotlin` here is only the Compose compiler plugin: there is no
org.jetbrains.kotlin.android in the build, AGP brings Kotlin itself, and the two
have to agree. That pairing is the thing to look at first if this one fails to
configure, because moving AGP moves the Kotlin underneath it.

okhttp and play-services-auth are the two majors. The uncertainty worth naming
was Coil, compiled against okhttp 4 and now resolving onto 5. Two things settled
it: R8 ran over Coil's bytecode with no okhttp or coil entry in
proguard-rules.pro and resolved every reference, and the failure mode would have
been every remote cover failing at once on the first screen rather than anything
subtle. Run on a device; the app works.

That bump surfaced 53 compiler warnings, all cleared here. None is a behaviour
change -- they are code the new compiler can now prove is dead.

  * 19 from okhttp 5, where Response.body became non-null. Where one of those
    guards carried a real branch -- the empty-body errors in AudioHttp, AppUpdate
    and ShareUtils -- the branch went with it, because it was unreachable, not
    because the case stopped mattering.
  * 32 from Kotlin 2.4, which follows nullability further than 2.2 did: through a
    boolean derived from a null check (DeezerAudioSource), into a lambda declared
    after the assignment it depends on (DjVoice -- and the lambda declared before
    it correctly keeps its safe call), and far enough out of an enclosing `if` to
    prove a `when` had a dead branch (DownloadsScreen).
  * 2 deprecations, handled differently because they are different things.
    LocationListener.onStatusChanged is suppressed on purpose: it has no default
    implementation before API 30 and minSdk is 29, so dropping the override is an
    AbstractMethodError on a device running 29. The media3 one is migrated --
    1.11 deprecated AcceptedResultBuilder(MediaSession) for
    AcceptedResultBuilder(MediaSession, ControllerInfo), and onConnect already had
    the controller in hand. Not a pure rename: the two-argument form seeds the
    *player* commands from the defaults for that particular controller rather than
    the session's generic ones, which is why the one-argument form went. The
    session commands are unaffected -- they are set explicitly on the next line,
    and that is what carries the custom "toggle saved" command the notification
    button needs.

53 warnings is enough noise to hide the next one that means something, and the
okhttp group was exactly the shape a real 5.x break would have taken.

The engine is untouched: not one line of Rust differs from 3.7.0, and there is no
version handshake between the two, so an old .so would behave identically. The
core version follows the app's as always, so it moves to 3.7.1 as a number only.

+++ app/src/main/java/com/varuna/rustify/metrics/ListeningTracker.kt
@@ -99,7 +99,7 @@ class ListeningTracker(private val appContext: Context) {
             // Rotation: keep the most recent ~5000 entries
             val trimmed = if (arr.length() > 5000) {
                 val fresh = JSONArray()
-                for (i in arr.length() - 5000 until arr.length()) fresh.put(arr.getJSONObject(i.toInt()))
+                for (i in arr.length() - 5000 until arr.length()) fresh.put(arr.getJSONObject(i))
                 fresh
             } else arr
             file.writeText(trimmed.toString())
